third_party/blink/web_tests/fast/js/bom-in-file-retains-correct-offset-expected.txt

This ensures that BOM's scattered through a source file do not break parsing

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyReparsed is true
PASS successfullyParsed is true

TEST COMPLETE
